Phase 2: Execution foundation — exchange clients, engine, Decimal migration
Add cross-platform arbitrage execution capability (disabled by default):

- Kalshi RSA-PSS authentication (execution/kalshi_auth.py)
- Kalshi trading client with order placement, status, cancellation (execution/kalshi_client.py)
- Polymarket py-clob-client SDK wrapper with GTC/FOK orders (execution/polymarket_client.py)
- Cross-platform execution engine with maker-first and parallel strategies (execution/engine.py)
- Execution data models using Decimal (execution/models.py)
- API endpoints: POST /execute, GET /execution/status
- Decimal migration: all financial arithmetic now uses decimal.Decimal
- decimal_utils.py for safe Decimal/JSON conversion
- 71 new execution tests, 224 total tests passing
- New dependencies: py-clob-client, cryptography

Safety: EXECUTION_ENABLED=false and EXECUTION_DRY_RUN=true by default.
Kalshi defaults to demo API (demo-api.kalshi.co).

+++ backend/binance.py
@@ -4,6 +4,7 @@ Single source of truth for BTC price fetching — used by both
 fetch_current_polymarket and fetch_current_kalshi.
 """
 import logging
+from decimal import Decimal
 from typing import Optional, Tuple
 import aiohttp
 from http_utils import fetch_json
@@ -12,17 +13,17 @@ from config import BINANCE_PRICE_URL, BINANCE_KLINES_URL, SYMBOL
 logger = logging.getLogger(__name__)
 
 
-async def get_binance_current_price(session: aiohttp.ClientSession) -> Tuple[Optional[float], Optional[str]]:
+async def get_binance_current_price(session: aiohttp.ClientSession) -> Tuple[Optional[Decimal], Optional[str]]:
     """Fetch current BTC spot price from Binance. Returns (price, error)."""
     try:
         data = await fetch_json(session, BINANCE_PRICE_URL, params={"symbol": SYMBOL})
-        return float(data["price"]), None
+        return Decimal(data["price"]), None
     except Exception as e:
         logger.error("Binance current price fetch failed: %s", e)
         return None, str(e)
 
 
-async def get_binance_open_price(session: aiohttp.ClientSession, target_time_utc) -> Tuple[Optional[float], Optional[str]]:
+async def get_binance_open_price(session: aiohttp.ClientSession, target_time_utc) -> Tuple[Optional[Decimal], Optional[str]]:
     """Fetch the 1-hour candle open price for a given UTC time. Returns (price, error)."""
     try:
         timestamp_ms = int(target_time_utc.timestamp() * 1000)
@@ -37,7 +38,7 @@ async def get_binance_open_price(session: aiohttp.ClientSession, target_time_utc
         if not data:
             return None, "Candle not found yet"
 
-        open_price = float(data[0][1])
+        open_price = Decimal(data[0][1])
         return open_price, None
     except Exception as e:
         logger.error("Binance open price fetch failed: %s", e)
